requirements.txt
playsound==1.2.2
pandas==1.5.3
Pillow==9.4.0
whisper==1.1.10
pysrt==1.1.2
googletrans==4.0.0rc1
Gooey==1.0.8.1
SpeechRecognition==3.10.0
nltk==3.6.2
torch==2.0.1+cu117
faster-whisper==0.7.1
translatepy==2.3
winshell==0.6
pytube==15.0.0